feature: add favicon and header for app with meta tags
- Remove comments in baseQueryWithReauth.ts
- Remove comments in authApi.ts
- Add comments in useSignIn.ts
- Move logic by check auth to index.tsx from _app.tsx
- Update layout.tsx: add checking for Storage.getToken()
- Changed the call sequence in useSidebar.ts for logout
